Added more resources and editor functionality

+++ Actions/Actions.MainForm.vb
@@ -165,8 +165,8 @@ Public Class Actions_MainForm
 
     Private Sub OpenFileDialog1_FileOk(sender As Object, e As System.ComponentModel.CancelEventArgs) Handles OpenFileDialog1.FileOk
         Scintilla1.Text = File.ReadAllText(OpenFileDialog1.FileName)
-        Text = "Actions - " & Path.GetFileName(OpenFileDialog1.FileName)
         ActionFile = OpenFileDialog1.FileName
+        Text = "Actions - " & Path.GetFileName(ActionFile)
     End Sub
 
     Private Sub ToolStripButton3_Click(sender As Object, e As EventArgs) Handles ToolStripButton3.Click
@@ -174,17 +174,41 @@ Public Class Actions_MainForm
     End Sub
 
     Private Sub ToolStripButton4_Click(sender As Object, e As EventArgs) Handles ToolStripButton4.Click
-        If SaveFileDialog1.ShowDialog() = Windows.Forms.DialogResult.OK Then
-            File.WriteAllText(SaveFileDialog1.FileName, Scintilla1.Text, ASCII)
-            ActionFile = SaveFileDialog1.FileName
+        If ActionFile Is Nothing Or Not File.Exists(ActionFile) Then
+            If SaveFileDialog1.ShowDialog() = Windows.Forms.DialogResult.OK Then
+                File.WriteAllText(SaveFileDialog1.FileName, Scintilla1.Text, ASCII)
+                ActionFile = SaveFileDialog1.FileName
+            End If
+        Else
+            File.WriteAllText(ActionFile, Scintilla1.Text, ASCII)
         End If
+        Text = "Actions - " & Path.GetFileName(ActionFile)
     End Sub
 
     Private Sub ToolStripButton1_Click(sender As Object, e As EventArgs) Handles ToolStripButton1.Click
+        If ActionFile Is Nothing Or Not File.Exists(ActionFile) Then
+            If SaveFileDialog1.ShowDialog() = Windows.Forms.DialogResult.OK Then
+                File.WriteAllText(SaveFileDialog1.FileName, Scintilla1.Text, ASCII)
+                ActionFile = SaveFileDialog1.FileName
+            Else
+                Exit Sub
+            End If
+            Text = "Actions - " & Path.GetFileName(ActionFile)
+        End If
         ProgressPanel.ActionFile = ActionFile
         ProgressPanel.ActionRunning = True
         ProgressPanel.IsInValidationMode = True
         WindowState = FormWindowState.Minimized
         ProgressPanel.ShowDialog(MainForm)
+    End Sub
+
+    Private Sub Scintilla1_TextChanged(sender As Object, e As EventArgs) Handles Scintilla1.TextChanged
+        If ActionFile IsNot Nothing And File.Exists(ActionFile) Then
+            If File.ReadAllText(ActionFile).ToString() = Scintilla1.Text Then
+                Text = "Actions - " & Path.GetFileName(ActionFile)
+            Else
+                Text = "Actions - " & Path.GetFileName(ActionFile) & "*"
+            End If
+        End If
     End Sub
 End Class
